requirements.txt
altair==5.5.0
anyio==4.8.0
ask-sdk-core==1.19.0
ask-sdk-model==1.82.0
ask-sdk-runtime==1.19.0
asttokens==2.4.1
async-timeout==5.0.1
asyncio==3.4.3
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.36.25
botocore==1.36.25
bs4==0.0.2
cachetools==5.5.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
exceptiongroup==1.2.1
executing==2.1.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.7
imageio==2.35.1
incremental==22.10.0
ipykernel==6.29.5
ipython==8.27.0
itemadapter==0.9.0
itemloaders==1.3.1
jedi==0.19.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.2.2
MarkupSafe==3.0.2
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
MouseInfo==0.1.3
mpmath==1.3.0
narwhals==1.34.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==2.0.1
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
parsel==1.9.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.2
pooch==1.8.2
prompt_toolkit==3.0.47
Protego==0.3.1
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pydeck==0.9.1
PyDispatcher==2.0.7
PyGetWindow==0.0.9
Pygments==2.18.0
PyMatting==1.1.12
PyMsgBox==1.0.9
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
pyshp==2.3.1
PySocks==1.7.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytweening==1.2.0
pytz==2024.1
pywin32==306
pyzmq==26.2.0
queuelib==1.7.0
redis==6.1.0
referencing==0.35.1
regex==2024.9.11
rembg==2.0.59
requests==2.32.3
requests-file==2.1.0
rpds-py==0.20.0
s3transfer==0.11.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
Scrapy==2.11.2
seaborn==0.13.2
selenium==4.22.0
service-identity==24.1.0
SHP==2.0.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
stats==0.1.2a0
streamlit==1.45.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tldextract==5.1.2
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.29.2
w3lib==2.2.1
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==6.4.post2
